Revert "Flip yinverted XML tag for analog sticks"
Y axis was inverted because some controllers on OSX were observed to have
inverted axes. The peripheral API was to be adjusted accordingly.

Instead, however, the peripheral API is remaining the same (up is 1.0, down
is -1.0) to conform to the XInput api. See https://github.com/garbear/xbmc/commit/05124db

This reverts commit 0798e8c13c44025534c93fae75d542fbfea3efba.