Fix loss of control if A becomes unmapped
When a feature is updated, it might conflict with an existing feature.
Instead of erasing the existing feature, an attempt is made to update its
primitives to the old feature being overwritten.

For example, if the button map is:

A -> 0
B -> 1

And the user maps "B -> 0", then the new button map will look like:

A -> 1
B -> 0

Here, instead of erasing the A (button 0) when it conflicts with the new
B (also button 0), the A is updated to point to B's old button (1).